requirements.txt

ccxt
pandas
numpy
backtrader
fastapi
uvicorn
python-dotenv
psycopg2-binary
redis
celery
